it just works (except admin) - todd howard

+++ internal/api/routes/guilds/delete.go
@@ -2,6 +2,7 @@ package guilds
 
 import (
 	"context"
+	"database/sql"
 	"fmt"
 	"net/http"
 	"os"
@@ -49,7 +50,7 @@ func deleteGuild(c *gin.Context) {
 
 	var isOwner bool
 	var isDm bool
-	if err := db.Db.QueryRow("SELECT EXISTS (SELECT 1 FROM userguilds WHERE guild_id=$1 AND user_id=$2 AND owner=true), EXISTS (SELECT 1 FROM guilds WHERE guild_id = $1 AND dm = true)", guildId, user.Id).Scan(&isOwner, &isDm); err != nil {
+	if err := db.Db.QueryRow("SELECT EXISTS (SELECT 1 FROM userguilds WHERE guild_id=$1 AND user_id=$2 AND owner=true), EXISTS (SELECT 1 FROM guilds WHERE id = $1 AND dm = true)", guildId, user.Id).Scan(&isOwner, &isDm); err != nil {
 		logger.Error.Println(err)
 		c.JSON(http.StatusInternalServerError, errors.Body{
 			Error:  err.Error(),
@@ -85,11 +86,19 @@ func deleteGuild(c *gin.Context) {
 		})
 		return
 	}
-	defer func() {
-		if err := tx.Rollback(); err != nil {
-			logger.Warn.Printf("unable to rollback error: %v\n", err)
-		}
-	}() //rollback changes if failed
+	defer tx.Rollback() //rollback changes if failed
+
+	var guildImageId int64
+	if err := tx.QueryRowContext(ctx, "SELECT id FROM files WHERE guild_id = $1", guildId).Scan(&guildImageId); err != nil && err != sql.ErrNoRows {
+		logger.Error.Println(err)
+		c.JSON(http.StatusInternalServerError, errors.Body{
+			Error:  err.Error(),
+			Status: errors.StatusInternalError,
+		})
+		return
+	} else if err == sql.ErrNoRows {
+		guildImageId = -1
+	}
 
 	fileIds, err := tx.QueryContext(ctx, `SELECT f.id FROM files f INNER JOIN msgs ON msgs.id = f.msg_id WHERE msgs.guild_id = $1`, guildId)
 	if err != nil {
@@ -100,17 +109,22 @@ func deleteGuild(c *gin.Context) {
 		})
 		return
 	}
-	defer fileIds.Close()
+	var fileIdsToDelete []int64
 
-	var guildImageId int64
-	if err := tx.QueryRowContext(ctx, "SELECT id FROM files WHERE guild_id = $1", guildId).Scan(&guildImageId); err != nil {
-		logger.Error.Println(err)
-		c.JSON(http.StatusInternalServerError, errors.Body{
-			Error:  err.Error(),
-			Status: errors.StatusInternalError,
-		})
-		return
+	for fileIds.Next() {
+		var fileId int64
+		if err := fileIds.Scan(&fileId); err != nil {
+			logger.Error.Println(err)
+			c.JSON(http.StatusInternalServerError, errors.Body{
+				Error:  err.Error(),
+				Status: errors.StatusInternalError,
+			})
+			return
+		}
+		fileIdsToDelete = append(fileIdsToDelete, fileId)
 	}
+
+	fileIds.Close()
 
 	if _, err := tx.ExecContext(ctx, "DELETE FROM guilds WHERE id=$1", guildId); err != nil {
 		logger.Error.Println(err)
@@ -140,21 +154,14 @@ func deleteGuild(c *gin.Context) {
 		return
 	}
 
-	if err := os.Remove(fmt.Sprintf("uploads/guild/%d.lz4", guildImageId)); err != nil {
-		logger.Warn.Printf("unable to remove file: %v\n", err)
+	if guildImageId != -1 {
+		if err := os.Remove(fmt.Sprintf("uploads/guild/%d.lz4", guildImageId)); err != nil {
+			logger.Warn.Printf("unable to remove file: %v\n", err)
+		}
 	}
 
-	for fileIds.Next() {
-		var fileId int64
-		if err := fileIds.Scan(&fileId); err != nil {
-			logger.Error.Println(err)
-			c.JSON(http.StatusInternalServerError, errors.Body{
-				Error:  err.Error(),
-				Status: errors.StatusInternalError,
-			})
-			return
-		}
-		if err := os.Remove(fmt.Sprintf("uploads/%d.lz4", fileId)); err != nil {
+	for _, fileId := range fileIdsToDelete {
+		if err := os.Remove(fmt.Sprintf("uploads/msg/%d.lz4", fileId)); err != nil {
 			logger.Warn.Printf("unable to remove file: %v\n", err)
 		}
 	}
